Corrected the Animal class constructor from int to double for the totalWeight parameter. The same correction is made in the BirdsCattle.cs, Chicken.cs, Duck.cs and Turkey.cs

+++ trunk/TheButcherCalulatorConsole/TheButcherCalculator/BirdsCattle.cs
@@ -35,7 +35,7 @@ namespace TheButcherCalculator
         }
 
         //constructor
-        protected BirdsCattle(string kind, int totalWeight)
+        protected BirdsCattle(string kind, double totalWeight)
             : base(kind, totalWeight)
         {
         }
